using res instead of httpservletres

+++ core/src/main/java/com/flipkart/gjex/core/context/AccessLogContext.java
@@ -18,7 +18,7 @@ import java.util.Map;
 public class AccessLogContext {
     String clientIp;
     String resourcePath;
-    Integer contentLength;
+    Long contentLength;
     Long responseTime;
     Integer responseStatus;
     @Builder.Default
